+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 18.09.2018
+
+* [Android] ToDo List: Adding LiveData
+
 #### 17.09.2018
 
 * [Android] ToDo List: Update Task